Backend/requirements.txt
amqp==5.3.1
asgiref==3.7.2
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.2
Django==4.2.10
django-cors-headers==4.7.0
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
kombu==5.5.4
msgpack==1.1.0
mysqlclient==2.2.7
packaging==25.0
paypalrestsdk==1.13.3
prompt_toolkit==3.0.51
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
redis==6.0.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.4.4
stripe==12.1.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
